+++ media/utils/addGHUserInfo.js
@@ -1,8 +1,12 @@
-const addGHUserInfo = ({login, avatar}) => {
-    $('.wm-logo').after(`
-    <span class='loginHolder'>
-        <img src="${avatar}" />
-        <p>Logged in as ${login}</p>
-    </span>`);
+const addGHUserInfo = ({ login, avatar, isStarred }) => {
+  $(".replace-user-img").replaceWith(`
+        <img src="${avatar}" class="login-image"/>
+    `);
+  $(".login-info").append(`
+        <p class="login-info-text">
+            Logged in as ${login}
+        </p>
+    `);
+  isStarred && $(".star-us-row").remove();
 };
 export default addGHUserInfo;