+++ media/utils/addPRsToDoc.js
@@ -5,6 +5,7 @@ import parseComments from "./parseComments.js";
 import addActionButtons from "./addActionButtons.js";
 
 const addPRsToDoc = (prs) => {
+  
   addActionButtons();
   $("#ghHolder").append(`
   <h3>Pull Requests</h3>
@@ -14,41 +15,46 @@ const addPRsToDoc = (prs) => {
     pr.comments.forEach((comment) => (mdComments += parseComments(comment)));
 
     $("#ghHolder").append(`
+    <div class="anim-fade-in">
       <details ${!index ? "open" : ""}>
         <summary class="pr-title">
-        <img class="pr-state" src="${
-          pr.state === "closed"
-            ? "https://raw.githubusercontent.com/primer/octicons/main/icons/git-merge-24.svg"
-            : "https://raw.githubusercontent.com/primer/octicons/main/icons/git-pull-request-24.svg"
-        }"
-        />
-        <a 
-       href="${pr.url}" target="_blank" title="View this PR on github">${
-      pr.title
-    }
-      </a>
+          <div>
+            <div class="details-state">
+              <div class="icon-holder">
+                <i class='codicon codicon-chevron-right'></i>
+                <i class='codicon codicon-chevron-down'></i>
+              </div>
+              <a 
+              href="${pr.url}" target="_blank" title="View this PR on github">${
+              pr.title
+              }</a>
+            </div>
+          </div>
+          <div class="icon-holder">${
+            pr.state === "closed"
+            ? "<i class='codicon codicon-git-merge'></i>"
+              : "<i class='codicon codicon-git-pull-request'></i>"
+          }
+          </div>
       </summary>
-        <div>
-        <div class="pr-header">
-          <div class="pr-owner">
+        <div class="Box">
+        <div class="Box-header d-flex">
             <p class="pr-poster" title="View this user on github">
               <a class="pr-author-combo" href="${
                 pr.userLink
               }"><img class='pr-author-img' src="${pr.userImage}" />${
-      pr.user
-    }</a>
+               pr.user
+              } </a>
             </p>
             <p class="pr-date">
-              ${dateToHumanReadable(pr.created_at)}
+                on ${dateToHumanReadable(pr.created_at)}
             </p>
-          </div>
         </div>
-          <div class="pr-body">
+          <div class="Box-body">
             ${
               pr?.body
-                ? replaceIssueLinks(
-                    replaceUserTags(marked.parse(pr.body)),
-                    pr.repo_url
+                ? (
+                    replaceUserTags(marked.parse(pr.body, { gfm: true, breaks: true }))
                   )
                 : ""
             }
@@ -56,6 +62,7 @@ const addPRsToDoc = (prs) => {
           ${mdComments}
         </div>
       </details>
+    </div>
       `);
   });
 };