+++ src/extension.ts
@@ -27,6 +27,7 @@ import multiSelectCommandHandler from "./utils/commands/multiSelect";
 import selectCommandHandler from "./utils/commands/select";
 import showCommandHandler from "./utils/commands/show";
 import debugLogger from "./utils/vscode/debugLogger";
+import checkIfUserStarred from "./utils/github/checkIfUserStarred";
 
 // repo information
 let owner: string | undefined = "";
@@ -86,7 +87,7 @@ export async function activate(context: vscode.ExtensionContext) {
   // create the hover provider
   let wmHover = hover({ reporter });
 
-  let repoInfo = await getRepoInfo();
+  let repoInfo = await getRepoInfo({});
   repo = repoInfo?.repo;
   owner = repoInfo?.owner;
   debugLogger(`repo: ${repo}`);
@@ -178,10 +179,12 @@ export async function activate(context: vscode.ExtensionContext) {
           error: issuesWithTitlesAndGroupedComments,
         });
       }
-
+      let sortedPRs = issuesWithTitlesAndGroupedComments?.sort(
+        (a: any, b: any) => b.comments.length - a.comments.length
+      );
       provider.sendMessage({
         command: "prs",
-        data: issuesWithTitlesAndGroupedComments,
+        data: sortedPRs,
       });
     } else {
       vscode.commands.executeCommand("watermelon.multiSelect");
@@ -214,9 +217,12 @@ export async function activate(context: vscode.ExtensionContext) {
           error: issuesWithTitlesAndGroupedComments,
         });
       }
+      let sortedPRs = issuesWithTitlesAndGroupedComments?.sort(
+        (a: any, b: any) => b.comments.length - a.comments.length
+      );
       provider.sendMessage({
         command: "prs",
-        data: issuesWithTitlesAndGroupedComments,
+        data: sortedPRs,
       });
     }
   };
@@ -244,7 +250,7 @@ export async function activate(context: vscode.ExtensionContext) {
     )
   );
 
-  vscode.authentication.getSession("github", []).then((session: any) => {
+  vscode.authentication.getSession("github", []).then(async (session: any) => {
     setLoggedIn(true);
     provider.sendMessage({
       command: "session",
@@ -252,6 +258,37 @@ export async function activate(context: vscode.ExtensionContext) {
       data: session.account.label,
     });
     debugLogger(`session: ${JSON.stringify(session)}`);
+    const credentials = new Credentials();
+    debugLogger(`got credentials`);
+    await credentials.initialize(context);
+    debugLogger("intialized credentials");
+    octokit = await credentials.getOctokit();
+    let githubUserInfo = await getGitHubUserInfo({ octokit });
+    debugLogger(`githubUserInfo: ${JSON.stringify(githubUserInfo)}`);
+    let username = githubUserInfo.login;
+    context.globalState.update("startupState", { username });
+    reporter?.sendTelemetryEvent("githubUserInfo", { username });
+    let isStarred = await checkIfUserStarred({ octokit });
+
+    provider.sendMessage({
+      command: "user",
+      data: {
+        login: githubUserInfo.login,
+        avatar: githubUserInfo.avatar_url,
+        isStarred,
+      },
+    });
+    let dailySummary = await getDailySummary({
+      octokit,
+      owner: owner || "",
+      repo: repo || "",
+      username: username || "",
+    });
+    debugLogger(`dailySummary: ${JSON.stringify(dailySummary)}`);
+    provider.sendMessage({
+      command: "dailySummary",
+      data: dailySummary,
+    });
   });
 
   vscode.window.onDidChangeTextEditorSelection(async (selection) => {